Adding template for batch trigger
